crawler/python/requirements.txt
pandas
beautifulsoup4
requests
lxml
ipython
selenium
pillow

